Watch packages for file changes and reload Meteor
This is done by merging the watchlists for the slices of the packages
and manually watching them.  When the watching finishes, a
`force.lazyreload` file is written to the disk, causing Meteor to rebundle
the application. During this rebundle, the `force.lazyreload` file is
deleted, such that it doesn't clutter up the filesystem.

Unfortunately, this means that the application will be rebundled twice,
which is slightly inefficient. Fortunately, for most applications this
isn't too big of an issue. If it turns out to be a big issue, I may add
a flag to the .lazypkgs.yml file allowing you to disable the cleanup.